+++ client/index.ts
@@ -1,5 +1,9 @@
-import validateConfig from '../packages/validate-config'
+import validateConfig from 'validate-config'
 
 export { ReferenceDataStore } from './src/reference-data'
 export * as definitionService from './src/definition-manipulation'
+export {
+  AutoSaveControlledForm,
+  useAutoSaveControlledForm,
+} from './src/components/AutoSaveControlledForm'
 export { validateConfig }