+++ CHANGELOG.md
@@ -12,3 +12,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - isElementVisble function
 - definition service
 - referenceDataStore class
+- AutoSaveControlledForm component